+++ website/src/types/utilities.ts
@@ -1,0 +1,7 @@
+import {DrawerUtilityButton} from 'src/types/drawer';
+
+export interface UtilityCategory {
+    name: string;
+    description: string;
+    buttons: DrawerUtilityButton[];
+}